WIP: 1/3 #185: SC_Plugin: Extract execution related methods to a new class and file.
No content changes are done in this commit in order to allow Git to better follow the change history.

main.ts is renamed temporarily in order to make Git notice that the content of the new file ShellCommandExecutor.ts is not actually new content, but originates from an old file.

main-RENAMED-TEMPORARILY.ts will be renamed back to main.ts in the next commit.