#178: Variable.getValue(): Return an object instead of a string.
The object contains the value (if getting one succeeded) or error messages (if getting a value failed).

Variable.getErrorMessages() is removed.

CustomVariable: Changed the class to define generateValue() instead of getValue, as then I don't have to make this class to create a similar object as Variable.getValue() does. generateValue() actually was defined before, but it was dummy and never called.