#185: ParsingProcess.processRest(): Fix incorrectly returning true when a previous parsing was unsuccessful and no more VariableSets were left.
In practise, this bug caused shell command execution _via menus_ to crash if the variables were parsed when displaying a menu title (but only when the variable parsing failed). The crash happened after clicking the menu item, when trying to execute a command because processRest() returned true and so allowed the program to proceed to call ShellCommandExecutor.executeShellCommand(), which received a null as a shell_command instead of a string.