#37: PromptModal.onOpen(): Use ShellCommandParsingProcess in shell command preview.
This makes the previewed variables a bit more accurate, when not all variables are parsed again by the modal.

Test e.g. with `echo {{date:ss}}`, without this commit, the prompt shell command preview will display seconds that update every time a PromptField's value changes (= after every keypress when typing text). Correct behaviour is that most variables are parsed before opening the prompt, and so the seconds should not change when typing values.