#145: OutputModal: Change hotkeys to use letters instead of numbers.
Also alter tooltip texts.

Also changes how output channels are included in the modal. If a channel does not define a hotkey, it's excluded. Earlier excluded channels were defined in OutputModal, not in the OutputChannelDriver subclasses. In practise, the same set of output channels is still included in the modal as before.