#37: Prompts: Fix autocomplete menu crashing sometimes.
The crash happened because a TextAreaComponent/TextComponent did not always exist right after calling .addTextArea()/.addText().

I also changed some autocomplete creations to happen in a .then() callback, to make the execution flow cleaner. I did not do this in PromptFieldModel because there the autocomplete creations are grouped together in a way that I did not want to change, but it can be done later if needed.