#240: In addition to increasing minAppVersion, also install newer Obsidian API version.
While upgrading the API version, TypeScript compiler started to give errors related to using objects that were created by `cloneObject()`. I'll also include fixes related to those errors in this commit. Not the cleanest way to commit, but so be it.

Upgrading Obsidian API to 0.5.x required to also install "@codemirror/state" and "@codemirror/view".