#242: Eslint: Do not check 'lib' folder.
'lib' contains code written by others (copied from the internet), and I want to minimize all changes to that kind of code. When there are fewer changes, it makes it easier to compare the code to the original source, if needed. Also bringing updates from the original source is easier.