#243/#76: createShellCommandField(): Fix references to SC_MainSettingsTab.
The mistake was made in commit 0dc54b2ab0361e8dd83c35a5e12647a7a7f332f8 / #76 when moving the logic to a separate file/function.

The mistake did not cause any problems in practise, because nothing actually used the reference. #243 needs to use it, so that's why I fixed it while doing #243.

I don't understand why TypeScript compiler does not yell about this kind of mistakes. When 'this' refers to a function, it's a completely different thing than a specific class 'SC_MainSettingsTab'.