#265: parseVariables(): Make the function asynchronous.
This allows creating new variables that depend on asynchronous methods, such as {{note_content}}.

I'm using a lot of resolve(null) calls in the Variable_* classes. Returning a 'null' value means that there was some problem. Normally reject() would be used here instead of resolve(), but the variable parsing system is currently built on the assumption that in case of a user-showable error, null is returned, instead of throwing anything. I want to change this in the future, so that errors will be reported either by throwing them or by using reject(). I'll need to learn what are the differences between those two.