#275: Make OutputChannels return Promises.
Makes it possible to improve output message queuing in "realtime" mode. It's not perfect, though.

I'm not certain if I have found all the spots where 'awaiting' needs to be done, or where a Promise needs to be returned.

Later, allows extending to do whatever else comes to mind to do after output has been handled. So far there has not been a possibility for that - but then again not need either.