#298: Migrations: Add "default_value" field to CustomVariables' configuration in settings file.
This is done in a dynamic way that does not mention the "default_value" property explicitly. In the future, the same logic can add other new properties to CustomVariables, too.

This is the first time the model system is used in migrations, so needed to move model introductions a bit earlier in SC_Plugin.onload().