WIP: #306: Shell: Support augmenting shell command contents before variable parsing.
If I ever get #306 solved, it might utilize this ability in a way that would convert linebreaks into something that works for CMD.EXE.

This commit can also be used if the augmenting functionality is needed for something else than #306. For now, I'm not going to merge it to main before there's a practical need for it.