Command palette: Fix a latent parsing bug caused by #265.
If parsing was very quick, it could have been that the parsing result was discarded (set to undefined). Not a huge problem, because it was used for command palette text, and discarding it would have cause it just to be parsed again when a command is executed. I've never noticed this in practise.

This fix is done while inspecting #312, although this is not directly related.