#302: Fix Alias field width by increasing CSS rules' specificity.
The :not() additions made in commit 16b62d4aa4b233e442c32d70055e803d01835e45 increased the specificity of that particular rule. That lead to tt accidentally overriding the rules made for .SC-no-description. The result was that 'Alias' became 50% wide, although it should be 100% wide.

I fixed the issue by adding .setting-item to the accidentally overridden rules, so that their specificity is now increased, too.