#297: CustomShell: Add 'shell_platform' setting.
The setting defines the values of 'directory_separator' and 'path_separator' in case they are set to fall back to the shell's platform choice. Later, 'shell_platform' can be used for other OS dependent decisions, too.

Note that I'll probably remove the 'directory_separator' and 'path_separator' settings in order to simplify CustomShell configuration, after which the settings will always base their values on the shell's operating system. The removal can be reverted in the future, if the settings are actually wanted.