#297: CustomShellSettingsModal: Fix saving supported_platforms.
platformId was incorrectly always 'win32', even when enabling/disabling Linux or macOS. The reason was that the for loop was using a let variable defined outside the loop. I changed it to const in the for () parenthesis. This led to TypeScript not liking the statement, so I did some "off-topic" changes and created a Map from PlatformNames, which works better in the for loop.