#297: CustomShellSettingsModal: Show a warning if binary path does not seem to exist.
Also removes binary file existence check from CustomShell.augmentSpawn() because it gave false error messages for relative shell binary paths that are available via the PATH environment variable, but not detectable as files in the current working directory.

The new warning text system in the settings modal should be a good-enough safeguard against broken shell binary paths, and it's also shown to the user  earlier, not late as when executing something.