#297: CustomShell: Parse shell command wrapper separately from the actual shell command.
This gives better access to the actual, unwrapped shell command content.

Before this, "The shell command is empty. :(" error message was _not_ shown for an empty shell command if a Shell had a wrapper.

There are also other places, where wrapped shell command content was used before, but where unwrapped is used now.