#365: Prompt fields: Remove surplus config properties when switching field type.
Implemented a function to remove unneeded configuration properties from the prompt field, preserving only those that are necessary. This was done to optimize and streamline the amount of properties stored in a configuration. A deep equality check has also been introduced to avoid removal of properties that although surplus, don't match default values.