#424: getFileYAMLValue(): Refactor ability to read multiple values.
This mostly overwrites changes done by @jvsteiner in commit fcf1ae5038555560d2d58057e75075a8d729a767 but will work as a base for a new variable {{yaml_values}} in a later commit. The original commit added multi-value support to {{yaml_value}}, but I will create a new variable with a plural name.

The aim is to keep the behavior of {{yaml_value}} unchanged.